pyproject.toml
[project]
name = "jmap-mcp"
version = "0.1.0"
description = "MCP server for Fastmail JMAP interaction"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "uvloop>=0.17.0",
    "jmapc>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
jmap-mcp-server = "jmap_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
